+++ server/graphql/Substance.ts
@@ -9,8 +9,8 @@ import {
   Int,
 } from "type-graphql";
 
-import Task from "../entity/Task";
-import Record from "../entity/Record";
+import Task from "../entities/Task";
+import Record from "../entities/Record";
 
 import { DifficultyLevel } from "./Public";
 import { ITask } from "./Task";
